Add files via upload
As CS50 Project0 assignment, I had to design a front-end for Google Search, Google Image Search, and Google Advanced Search.
here are the specification;
https://cs50.harvard.edu/web/2020/projects/0/search/

URL of a short video demonstrating my implementation of Project 0: Search
This video show my application live, in action, and demonstrate that every feature required by the specification has been implemented;
https://www.youtube.com/watch?v=PPuHjUaOIWU

My website must meet the following requirements;

1-Pages. Your website should have at least three pages: one for Google Search, one for Google Image Search, and one for Google Advanced Search.
On the Google Search page, there should be links in the upper-right of the page to go to Image Search or Advanced Search. On each of the other two pages, there should be a link in the upper-right to go back to Google Search.

2-Query Text. On the Google Search page, the user should be able to type in a query, click “Google Search”, and be taken to the Google search results for that page.
Like Google’s own, your search bar should be centered with rounded corners. The search button should also be centered, and should be beneath the search bar.

3-Query Images. On the Google Image Search page, the user should be able to type in a query, click a search button, and be taken to the Google Image search results for that page.

4-Query Advanced. On the Google Advanced Search page, the user should be able to provide input for the following four fields (taken from Google’s own advanced search options)
Find pages with… “all these words:”
Find pages with… “this exact word or phrase:”
Find pages with… “any of these words:”
Find pages with… “none of these words:”

5-Appearance. Like Google’s own Advanced Search page, the four options should be stacked vertically, and all of the text fields should be left aligned.
Consistent with Google’s own CSS, the “Advanced Search” button should be blue with white text. When the “Advanced Search” button is clicked, the user should be taken to search results page for their given query.

6-Lucky. Add an “I’m Feeling Lucky” button to the main Google Search page. Consistent with Google’s own behavior, clicking this link should take users directly to the first Google search result for the query, bypassing the normal results page.

7-Aesthetics. The CSS you write should match Google’s own aesthetics as best as possible.